refactor(webchat): extract countdown timer logic into useInteractionTimeout hook
