+++ webchat/components/assistant-ui/tools/QuestionResponseCard.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { motion } from "framer-motion";
 import { useTranslation } from "react-i18next";
 import type { InteractionState, InteractionStatus } from "@/lib/adapters/hotplex-runtime-adapter";
+import { useInteractionTimeout } from "@/hooks/useInteractionTimeout";
 
 interface QuestionItem {
   id: string;
@@ -32,7 +33,7 @@ export function QuestionResponseCard({
   onToggle,
 }: QuestionResponseCardProps) {
   const { t } = useTranslation("chat");
-  const [timeLeft, setTimeLeft] = useState<number | null>(null);
+  const { timeLeft, activeStatus } = useInteractionTimeout(initialStatus, interactionState?.expiresAt);
 
   // Local state for answers: questionId -> value (or comma-separated values for multi-select)
   const [answers, setAnswers] = useState<Record<string, string>>({});
@@ -40,23 +41,6 @@ export function QuestionResponseCard({
   const [selectedMulti, setSelectedMulti] = useState<Record<string, Set<string>>>({});
   // Local state for custom text inputs: questionId -> string
   const [textAnswers, setTextAnswers] = useState<Record<string, string>>({});
-
-  useEffect(() => {
-    if (initialStatus !== "pending" || !interactionState?.expiresAt) {
-      setTimeLeft(null);
-      return;
-    }
-    const updateTime = () => {
-      const diff = Math.max(0, Math.floor((interactionState.expiresAt! - Date.now()) / 1000));
-      setTimeLeft(diff);
-    };
-    updateTime();
-    const interval = setInterval(updateTime, 1000);
-    return () => clearInterval(interval);
-  }, [initialStatus, interactionState?.expiresAt]);
-
-  const activeStatus: InteractionStatus =
-    initialStatus === "pending" && timeLeft !== null && timeLeft <= 0 ? "expired" : initialStatus;
 
   const isInteractive = activeStatus === "pending" || activeStatus === "failed";
